Doc: Fix QByteArray::toFloat snippet
Amends dc133765ec47e9625c49701f0ffd762b0ee1ad48

+++ src/corelib/doc/snippets/code/src_corelib_tools_qbytearray.cpp
@@ -363,7 +363,7 @@ a = str.toDouble(&ok);             // a == 0, ok == false
 //! [38float]
 QByteArray string("1234.56");
 bool ok;
-double a = string.toFloat(&ok);    // a == 1234.56, ok == true
+float a = string.toFloat(&ok);    // a == 1234.56, ok == true
 
 string = "1234.56 Volt";
 a = str.toFloat(&ok);              // a == 0, ok == false
